@@ -2,8 +2,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v3.19.6
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.0
 // source: internal/protos/exam.proto
 
 package exam
@@ -493,6 +493,61 @@ func (x *Projects) GetProjects() []*Project {
 	return nil
 }
 
+type ManagerProjects struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	ManagerName  string `protobuf:"bytes,1,opt,name=manager_name,json=managerName,proto3" json:"manager_name,omitempty"`
+	ProjectCount int64  `protobuf:"varint,2,opt,name=project_count,json=projectCount,proto3" json:"project_count,omitempty"`
+}
+
+func (x *ManagerProjects) Reset() {
+	*x = ManagerProjects{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_internal_protos_exam_proto_msgTypes[9]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ManagerProjects) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ManagerProjects) ProtoMessage() {}
+
+func (x *ManagerProjects) ProtoReflect() protoreflect.Message {
+	mi := &file_internal_protos_exam_proto_msgTypes[9]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ManagerProjects.ProtoReflect.Descriptor instead.
+func (*ManagerProjects) Descriptor() ([]byte, []int) {
+	return file_internal_protos_exam_proto_rawDescGZIP(), []int{9}
+}
+
+func (x *ManagerProjects) GetManagerName() string {
+	if x != nil {
+		return x.ManagerName
+	}
+	return ""
+}
+
+func (x *ManagerProjects) GetProjectCount() int64 {
+	if x != nil {
+		return x.ProjectCount
+	}
+	return 0
+}
+
 var File_internal_protos_exam_proto protoreflect.FileDescriptor
 
 var file_internal_protos_exam_proto_rawDesc = []byte{
@@ -531,36 +586,47 @@ var file_internal_protos_exam_proto_rawDesc = []byte{
 	0x63, 0x74, 0x73, 0x12, 0x32, 0x0a, 0x08, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x73, 0x18,
 	0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79,
 	0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x52, 0x08, 0x50,
-	0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x73, 0x32, 0x90, 0x03, 0x0a, 0x04, 0x45, 0x78, 0x61, 0x6d,
-	0x12, 0x3d, 0x0a, 0x0f, 0x47, 0x65, 0x74, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x42,
-	0x79, 0x49, 0x44, 0x12, 0x11, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65,
-	0x78, 0x61, 0x6d, 0x2e, 0x49, 0x44, 0x1a, 0x17, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74,
-	0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x12,
-	0x41, 0x0a, 0x0f, 0x47, 0x65, 0x74, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x4c, 0x69,
-	0x73, 0x74, 0x12, 0x14, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78,
-	0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x18, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72,
-	0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65,
-	0x65, 0x73, 0x12, 0x41, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x44, 0x65, 0x70, 0x61, 0x72, 0x74, 0x6d,
-	0x65, 0x6e, 0x74, 0x42, 0x79, 0x49, 0x44, 0x12, 0x11, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69,
-	0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x49, 0x44, 0x1a, 0x19, 0x2e, 0x70, 0x72, 0x69,
-	0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x44, 0x65, 0x70, 0x61, 0x72,
-	0x74, 0x6d, 0x65, 0x6e, 0x74, 0x12, 0x45, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x44, 0x65, 0x70, 0x61,
-	0x72, 0x74, 0x6d, 0x65, 0x6e, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x14, 0x2e, 0x70, 0x72, 0x69,
-	0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79,
-	0x1a, 0x1a, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d,
-	0x2e, 0x44, 0x65, 0x70, 0x61, 0x72, 0x74, 0x6d, 0x65, 0x6e, 0x74, 0x73, 0x12, 0x3b, 0x0a, 0x0e,
-	0x47, 0x65, 0x74, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x42, 0x79, 0x49, 0x44, 0x12, 0x11,
+	0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x73, 0x22, 0x59, 0x0a, 0x0f, 0x4d, 0x61, 0x6e, 0x61, 0x67,
+	0x65, 0x72, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x73, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x61,
+	0x6e, 0x61, 0x67, 0x65, 0x72, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x0b, 0x6d, 0x61, 0x6e, 0x61, 0x67, 0x65, 0x72, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x23, 0x0a,
+	0x0d, 0x70, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x02,
+	0x20, 0x01, 0x28, 0x03, 0x52, 0x0c, 0x70, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x43, 0x6f, 0x75,
+	0x6e, 0x74, 0x32, 0xde, 0x03, 0x0a, 0x04, 0x45, 0x78, 0x61, 0x6d, 0x12, 0x3d, 0x0a, 0x0f, 0x47,
+	0x65, 0x74, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x42, 0x79, 0x49, 0x44, 0x12, 0x11,
 	0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x49,
-	0x44, 0x1a, 0x16, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61,
-	0x6d, 0x2e, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x12, 0x3f, 0x0a, 0x0e, 0x47, 0x65, 0x74,
-	0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x14, 0x2e, 0x70, 0x72,
+	0x44, 0x1a, 0x17, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61,
+	0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x12, 0x41, 0x0a, 0x0f, 0x47, 0x65,
+	0x74, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x14, 0x2e,
+	0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d,
+	0x70, 0x74, 0x79, 0x1a, 0x18, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65,
+	0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x6c, 0x6f, 0x79, 0x65, 0x65, 0x73, 0x12, 0x41, 0x0a,
+	0x11, 0x47, 0x65, 0x74, 0x44, 0x65, 0x70, 0x61, 0x72, 0x74, 0x6d, 0x65, 0x6e, 0x74, 0x42, 0x79,
+	0x49, 0x44, 0x12, 0x11, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78,
+	0x61, 0x6d, 0x2e, 0x49, 0x44, 0x1a, 0x19, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79,
+	0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x44, 0x65, 0x70, 0x61, 0x72, 0x74, 0x6d, 0x65, 0x6e, 0x74,
+	0x12, 0x45, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x44, 0x65, 0x70, 0x61, 0x72, 0x74, 0x6d, 0x65, 0x6e,
+	0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x14, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79,
+	0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x1a, 0x2e, 0x70, 0x72,
+	0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x44, 0x65, 0x70, 0x61,
+	0x72, 0x74, 0x6d, 0x65, 0x6e, 0x74, 0x73, 0x12, 0x3b, 0x0a, 0x0e, 0x47, 0x65, 0x74, 0x50, 0x72,
+	0x6f, 0x6a, 0x65, 0x63, 0x74, 0x42, 0x79, 0x49, 0x44, 0x12, 0x11, 0x2e, 0x70, 0x72, 0x69, 0x6f,
+	0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x49, 0x44, 0x1a, 0x16, 0x2e, 0x70,
+	0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x50, 0x72, 0x6f,
+	0x6a, 0x65, 0x63, 0x74, 0x12, 0x3f, 0x0a, 0x0e, 0x47, 0x65, 0x74, 0x50, 0x72, 0x6f, 0x6a, 0x65,
+	0x63, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x14, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74,
+	0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x17, 0x2e, 0x70,
+	0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x50, 0x72, 0x6f,
+	0x6a, 0x65, 0x63, 0x74, 0x73, 0x12, 0x4c, 0x0a, 0x12, 0x47, 0x65, 0x74, 0x4d, 0x61, 0x6e, 0x61,
+	0x67, 0x65, 0x72, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x73, 0x12, 0x14, 0x2e, 0x70, 0x72,
 	0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61, 0x6d, 0x2e, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x1a, 0x17, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61,
-	0x6d, 0x2e, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74, 0x73, 0x42, 0x38, 0x5a, 0x36, 0x67, 0x69,
-	0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74,
-	0x79, 0x2d, 0x69, 0x6e, 0x66, 0x72, 0x61, 0x2f, 0x67, 0x6f, 0x5f, 0x65, 0x78, 0x61, 0x6d, 0x2f,
-	0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2f,
-	0x65, 0x78, 0x61, 0x6d, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x79, 0x1a, 0x1e, 0x2e, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2e, 0x65, 0x78, 0x61,
+	0x6d, 0x2e, 0x4d, 0x61, 0x6e, 0x61, 0x67, 0x65, 0x72, 0x50, 0x72, 0x6f, 0x6a, 0x65, 0x63, 0x74,
+	0x73, 0x30, 0x01, 0x42, 0x38, 0x5a, 0x36, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f,
+	0x6d, 0x2f, 0x70, 0x72, 0x69, 0x6f, 0x72, 0x69, 0x74, 0x79, 0x2d, 0x69, 0x6e, 0x66, 0x72, 0x61,
+	0x2f, 0x67, 0x6f, 0x5f, 0x65, 0x78, 0x61, 0x6d, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61,
+	0x6c, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2f, 0x65, 0x78, 0x61, 0x6d, 0x62, 0x06, 0x70,
+	0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -575,39 +641,42 @@ func file_internal_protos_exam_proto_rawDescGZIP() []byte {
 	return file_internal_protos_exam_proto_rawDescData
 }
 
-var file_internal_protos_exam_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
-var file_internal_protos_exam_proto_goTypes = []interface{}{
-	(*String)(nil),      // 0: priority.exam.String
-	(*ID)(nil),          // 1: priority.exam.ID
-	(*Empty)(nil),       // 2: priority.exam.Empty
-	(*Employee)(nil),    // 3: priority.exam.Employee
-	(*Employees)(nil),   // 4: priority.exam.Employees
-	(*Department)(nil),  // 5: priority.exam.Department
-	(*Departments)(nil), // 6: priority.exam.Departments
-	(*Project)(nil),     // 7: priority.exam.Project
-	(*Projects)(nil),    // 8: priority.exam.Projects
+var file_internal_protos_exam_proto_msgTypes = make([]protoimpl.MessageInfo, 10)
+var file_internal_protos_exam_proto_goTypes = []any{
+	(*String)(nil),          // 0: priority.exam.String
+	(*ID)(nil),              // 1: priority.exam.ID
+	(*Empty)(nil),           // 2: priority.exam.Empty
+	(*Employee)(nil),        // 3: priority.exam.Employee
+	(*Employees)(nil),       // 4: priority.exam.Employees
+	(*Department)(nil),      // 5: priority.exam.Department
+	(*Departments)(nil),     // 6: priority.exam.Departments
+	(*Project)(nil),         // 7: priority.exam.Project
+	(*Projects)(nil),        // 8: priority.exam.Projects
+	(*ManagerProjects)(nil), // 9: priority.exam.ManagerProjects
 }
 var file_internal_protos_exam_proto_depIdxs = []int32{
-	3, // 0: priority.exam.Employees.Employees:type_name -> priority.exam.Employee
-	5, // 1: priority.exam.Departments.Departments:type_name -> priority.exam.Department
-	7, // 2: priority.exam.Projects.Projects:type_name -> priority.exam.Project
-	1, // 3: priority.exam.Exam.GetEmployeeByID:input_type -> priority.exam.ID
-	2, // 4: priority.exam.Exam.GetEmployeeList:input_type -> priority.exam.Empty
-	1, // 5: priority.exam.Exam.GetDepartmentByID:input_type -> priority.exam.ID
-	2, // 6: priority.exam.Exam.GetDepartmentList:input_type -> priority.exam.Empty
-	1, // 7: priority.exam.Exam.GetProjectByID:input_type -> priority.exam.ID
-	2, // 8: priority.exam.Exam.GetProjectList:input_type -> priority.exam.Empty
-	3, // 9: priority.exam.Exam.GetEmployeeByID:output_type -> priority.exam.Employee
-	4, // 10: priority.exam.Exam.GetEmployeeList:output_type -> priority.exam.Employees
-	5, // 11: priority.exam.Exam.GetDepartmentByID:output_type -> priority.exam.Department
-	6, // 12: priority.exam.Exam.GetDepartmentList:output_type -> priority.exam.Departments
-	7, // 13: priority.exam.Exam.GetProjectByID:output_type -> priority.exam.Project
-	8, // 14: priority.exam.Exam.GetProjectList:output_type -> priority.exam.Projects
-	9, // [9:15] is the sub-list for method output_type
-	3, // [3:9] is the sub-list for method input_type
-	3, // [3:3] is the sub-list for extension type_name
-	3, // [3:3] is the sub-list for extension extendee
-	0, // [0:3] is the sub-list for field type_name
+	3,  // 0: priority.exam.Employees.Employees:type_name -> priority.exam.Employee
+	5,  // 1: priority.exam.Departments.Departments:type_name -> priority.exam.Department
+	7,  // 2: priority.exam.Projects.Projects:type_name -> priority.exam.Project
+	1,  // 3: priority.exam.Exam.GetEmployeeByID:input_type -> priority.exam.ID
+	2,  // 4: priority.exam.Exam.GetEmployeeList:input_type -> priority.exam.Empty
+	1,  // 5: priority.exam.Exam.GetDepartmentByID:input_type -> priority.exam.ID
+	2,  // 6: priority.exam.Exam.GetDepartmentList:input_type -> priority.exam.Empty
+	1,  // 7: priority.exam.Exam.GetProjectByID:input_type -> priority.exam.ID
+	2,  // 8: priority.exam.Exam.GetProjectList:input_type -> priority.exam.Empty
+	2,  // 9: priority.exam.Exam.GetManagerProjects:input_type -> priority.exam.Empty
+	3,  // 10: priority.exam.Exam.GetEmployeeByID:output_type -> priority.exam.Employee
+	4,  // 11: priority.exam.Exam.GetEmployeeList:output_type -> priority.exam.Employees
+	5,  // 12: priority.exam.Exam.GetDepartmentByID:output_type -> priority.exam.Department
+	6,  // 13: priority.exam.Exam.GetDepartmentList:output_type -> priority.exam.Departments
+	7,  // 14: priority.exam.Exam.GetProjectByID:output_type -> priority.exam.Project
+	8,  // 15: priority.exam.Exam.GetProjectList:output_type -> priority.exam.Projects
+	9,  // 16: priority.exam.Exam.GetManagerProjects:output_type -> priority.exam.ManagerProjects
+	10, // [10:17] is the sub-list for method output_type
+	3,  // [3:10] is the sub-list for method input_type
+	3,  // [3:3] is the sub-list for extension type_name
+	3,  // [3:3] is the sub-list for extension extendee
+	0,  // [0:3] is the sub-list for field type_name
 }
 
 func init() { file_internal_protos_exam_proto_init() }
@@ -616,7 +685,7 @@ func file_internal_protos_exam_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_internal_protos_exam_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*String); i {
 			case 0:
 				return &v.state
@@ -628,7 +697,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ID); i {
 			case 0:
 				return &v.state
@@ -640,7 +709,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Empty); i {
 			case 0:
 				return &v.state
@@ -652,7 +721,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*Employee); i {
 			case 0:
 				return &v.state
@@ -664,7 +733,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*Employees); i {
 			case 0:
 				return &v.state
@@ -676,7 +745,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*Department); i {
 			case 0:
 				return &v.state
@@ -688,7 +757,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*Departments); i {
 			case 0:
 				return &v.state
@@ -700,7 +769,7 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*Project); i {
 			case 0:
 				return &v.state
@@ -712,8 +781,20 @@ func file_internal_protos_exam_proto_init() {
 				return nil
 			}
 		}
-		file_internal_protos_exam_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_internal_protos_exam_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*Projects); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_internal_protos_exam_proto_msgTypes[9].Exporter = func(v any, i int) any {
+			switch v := v.(*ManagerProjects); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -731,7 +812,7 @@ func file_internal_protos_exam_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_internal_protos_exam_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   9,
+			NumMessages:   10,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
