@@ -2,8 +2,8 @@
 
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.19.6
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.0
 // source: internal/protos/exam.proto
 
 package exam
@@ -17,16 +17,17 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	Exam_GetEmployeeByID_FullMethodName   = "/priority.exam.Exam/GetEmployeeByID"
-	Exam_GetEmployeeList_FullMethodName   = "/priority.exam.Exam/GetEmployeeList"
-	Exam_GetDepartmentByID_FullMethodName = "/priority.exam.Exam/GetDepartmentByID"
-	Exam_GetDepartmentList_FullMethodName = "/priority.exam.Exam/GetDepartmentList"
-	Exam_GetProjectByID_FullMethodName    = "/priority.exam.Exam/GetProjectByID"
-	Exam_GetProjectList_FullMethodName    = "/priority.exam.Exam/GetProjectList"
+	Exam_GetEmployeeByID_FullMethodName    = "/priority.exam.Exam/GetEmployeeByID"
+	Exam_GetEmployeeList_FullMethodName    = "/priority.exam.Exam/GetEmployeeList"
+	Exam_GetDepartmentByID_FullMethodName  = "/priority.exam.Exam/GetDepartmentByID"
+	Exam_GetDepartmentList_FullMethodName  = "/priority.exam.Exam/GetDepartmentList"
+	Exam_GetProjectByID_FullMethodName     = "/priority.exam.Exam/GetProjectByID"
+	Exam_GetProjectList_FullMethodName     = "/priority.exam.Exam/GetProjectList"
+	Exam_GetManagerProjects_FullMethodName = "/priority.exam.Exam/GetManagerProjects"
 )
 
 // ExamClient is the client API for Exam service.
@@ -39,6 +40,7 @@ type ExamClient interface {
 	GetDepartmentList(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Departments, error)
 	GetProjectByID(ctx context.Context, in *ID, opts ...grpc.CallOption) (*Project, error)
 	GetProjectList(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Projects, error)
+	GetManagerProjects(ctx context.Context, in *Empty, opts ...grpc.CallOption) (grpc.ServerStreamingClient[ManagerProjects], error)
 }
 
 type examClient struct {
@@ -50,8 +52,9 @@ func NewExamClient(cc grpc.ClientConnInterface) ExamClient {
 }
 
 func (c *examClient) GetEmployeeByID(ctx context.Context, in *ID, opts ...grpc.CallOption) (*Employee, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Employee)
-	err := c.cc.Invoke(ctx, Exam_GetEmployeeByID_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Exam_GetEmployeeByID_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,8 +62,9 @@ func (c *examClient) GetEmployeeByID(ctx context.Context, in *ID, opts ...grpc.C
 }
 
 func (c *examClient) GetEmployeeList(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Employees, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Employees)
-	err := c.cc.Invoke(ctx, Exam_GetEmployeeList_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Exam_GetEmployeeList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -68,8 +72,9 @@ func (c *examClient) GetEmployeeList(ctx context.Context, in *Empty, opts ...grp
 }
 
 func (c *examClient) GetDepartmentByID(ctx context.Context, in *ID, opts ...grpc.CallOption) (*Department, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Department)
-	err := c.cc.Invoke(ctx, Exam_GetDepartmentByID_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Exam_GetDepartmentByID_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -77,8 +82,9 @@ func (c *examClient) GetDepartmentByID(ctx context.Context, in *ID, opts ...grpc
 }
 
 func (c *examClient) GetDepartmentList(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Departments, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Departments)
-	err := c.cc.Invoke(ctx, Exam_GetDepartmentList_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Exam_GetDepartmentList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -86,8 +92,9 @@ func (c *examClient) GetDepartmentList(ctx context.Context, in *Empty, opts ...g
 }
 
 func (c *examClient) GetProjectByID(ctx context.Context, in *ID, opts ...grpc.CallOption) (*Project, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Project)
-	err := c.cc.Invoke(ctx, Exam_GetProjectByID_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Exam_GetProjectByID_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -95,17 +102,37 @@ func (c *examClient) GetProjectByID(ctx context.Context, in *ID, opts ...grpc.Ca
 }
 
 func (c *examClient) GetProjectList(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Projects, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Projects)
-	err := c.cc.Invoke(ctx, Exam_GetProjectList_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Exam_GetProjectList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
+func (c *examClient) GetManagerProjects(ctx context.Context, in *Empty, opts ...grpc.CallOption) (grpc.ServerStreamingClient[ManagerProjects], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Exam_ServiceDesc.Streams[0], Exam_GetManagerProjects_FullMethodName, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	x := &grpc.GenericClientStream[Empty, ManagerProjects]{ClientStream: stream}
+	if err := x.ClientStream.SendMsg(in); err != nil {
+		return nil, err
+	}
+	if err := x.ClientStream.CloseSend(); err != nil {
+		return nil, err
+	}
+	return x, nil
+}
+
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Exam_GetManagerProjectsClient = grpc.ServerStreamingClient[ManagerProjects]
+
 // ExamServer is the server API for Exam service.
 // All implementations must embed UnimplementedExamServer
-// for forward compatibility
+// for forward compatibility.
 type ExamServer interface {
 	GetEmployeeByID(context.Context, *ID) (*Employee, error)
 	GetEmployeeList(context.Context, *Empty) (*Employees, error)
@@ -113,12 +140,16 @@ type ExamServer interface {
 	GetDepartmentList(context.Context, *Empty) (*Departments, error)
 	GetProjectByID(context.Context, *ID) (*Project, error)
 	GetProjectList(context.Context, *Empty) (*Projects, error)
+	GetManagerProjects(*Empty, grpc.ServerStreamingServer[ManagerProjects]) error
 	mustEmbedUnimplementedExamServer()
 }
 
-// UnimplementedExamServer must be embedded to have forward compatible implementations.
-type UnimplementedExamServer struct {
-}
+// UnimplementedExamServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedExamServer struct{}
 
 func (UnimplementedExamServer) GetEmployeeByID(context.Context, *ID) (*Employee, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetEmployeeByID not implemented")
@@ -138,7 +169,11 @@ func (UnimplementedExamServer) GetProjectByID(context.Context, *ID) (*Project, e
 func (UnimplementedExamServer) GetProjectList(context.Context, *Empty) (*Projects, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetProjectList not implemented")
 }
+func (UnimplementedExamServer) GetManagerProjects(*Empty, grpc.ServerStreamingServer[ManagerProjects]) error {
+	return status.Errorf(codes.Unimplemented, "method GetManagerProjects not implemented")
+}
 func (UnimplementedExamServer) mustEmbedUnimplementedExamServer() {}
+func (UnimplementedExamServer) testEmbeddedByValue()              {}
 
 // UnsafeExamServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ExamServer will
@@ -148,6 +183,13 @@ type UnsafeExamServer interface {
 }
 
 func RegisterExamServer(s grpc.ServiceRegistrar, srv ExamServer) {
+	// If the following call pancis, it indicates UnimplementedExamServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Exam_ServiceDesc, srv)
 }
 
@@ -259,6 +301,17 @@ func _Exam_GetProjectList_Handler(srv interface{}, ctx context.Context, dec func
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Exam_GetManagerProjects_Handler(srv interface{}, stream grpc.ServerStream) error {
+	m := new(Empty)
+	if err := stream.RecvMsg(m); err != nil {
+		return err
+	}
+	return srv.(ExamServer).GetManagerProjects(m, &grpc.GenericServerStream[Empty, ManagerProjects]{ServerStream: stream})
+}
+
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Exam_GetManagerProjectsServer = grpc.ServerStreamingServer[ManagerProjects]
+
 // Exam_ServiceDesc is the grpc.ServiceDesc for Exam service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -291,6 +344,12 @@ var Exam_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _Exam_GetProjectList_Handler,
 		},
 	},
-	Streams:  []grpc.StreamDesc{},
+	Streams: []grpc.StreamDesc{
+		{
+			StreamName:    "GetManagerProjects",
+			Handler:       _Exam_GetManagerProjects_Handler,
+			ServerStreams: true,
+		},
+	},
 	Metadata: "internal/protos/exam.proto",
 }
